Detect Ubuntu packages already installed
Avoid installing them again when it is not required.

Set up an extra job that uses Travis addon 'apt'
to pre-install the required dependencies.

+++ install.sh
@@ -9,8 +9,13 @@ PACKAGE_DEPENDENCIES="libaio1 rpm"
 
 cd "$(dirname "$(readlink -f "$0")")"
 
-sudo apt-get -qq update
-sudo apt-get --no-install-recommends -qq install $PACKAGE_DEPENDENCIES
+if dpkg -s $PACKAGE_DEPENDENCIES >/dev/null 2>/dev/null; then
+  echo "Oracle XE dependencies are already installed: $PACKAGE_DEPENDENCIES"
+else
+  echo "Installing Oracle XE dependencies: $PACKAGE_DEPENDENCIES"
+  sudo apt-get -qq update
+  sudo apt-get --no-install-recommends -qq install $PACKAGE_DEPENDENCIES
+fi
 
 df -B1 /dev/shm | awk 'END { if ($1 != "shmfs" && $1 != "tmpfs" || $2 < 2147483648) exit 1 }' ||
   ( sudo rm -r /dev/shm && sudo mkdir /dev/shm && sudo mount -t tmpfs shmfs -o size=2G /dev/shm )
